FIX: append os.Environ to docker build and docker run calls (#18)
* FIX: append os.Environ to docker build and docker run calls
* Add up, down, rm aliases - up, down, rm are natural to be in parallel with docker commands

Apply feedback including:

Linting, use of temp dir, error handling, normalizing naming standard, and context passing

---
# Context passing

https://go.dev/wiki/CodeReviewComments#contexts

Do not hold context in structs.

As best effort as I can get, I'm attempting to pass context into Run functions
called by kong. It seems like Kong Client is always first value in the binding
after a Bind() or BindTo() is called. The method signature to read as
Run(context.Context, *Cli) is difficult to set up properly, as such I have left
the context as the second param for kong commands.

See also:
https://github.com/alecthomas/kong/issues/48
https://github.com/alecthomas/kong/issues/144

---
Writing bytes directly to output

Per Saj's feedback:

It would be more efficient to restructure the program so that it need not buffer all of the output. Let the operating system do all of the work. This is especially true for large amounts of data, as may be the case here.

The io.Reader and io.Writer interfaces in Go are quite powerful, and should form the basis of your internal I/O design.

Conversions of string → []byte (and vice versa) require a memory allocation. Allocations are sometimes unavoidable, but they generate garbage. The garbage must be collected, later. In real programs, it is usually better to avoid unnecessary allocations and alleviate pressure on the garbage collector. If you already have []byte slice, use the []byte slice as-is.

The fmt package is useful when you truly need a formatted string. In all other cases, all it does is make your program slower. (fmt can only work with run-time reflection.)