Add build mount support (#20)
* make volumes available to read in build step

Add an potion allowing for an external cache to be mounted at build time.

Note: no data is saved to the host after this step.

The cache will need to be updated outside of the docker build step (by external
tooling) if desired.

see bind mounts in dockerfile source:
https://docs.docker.com/reference/dockerfile#run---mounttypebind

Cache mounts would seems like the most dockery-approved way to go
with these sorts of builds and features, but it doesn't work.

We want to build with no cache (but the cache mount) and currently there is no
way of doing that in a docker build.

--no-cache option disables ALL caches - build cache and mount caches.

To retain and use a cache under --no-cache when we just want to disable
build cache, we have no choice but use bind mounts as caches instead.

add volume mounts on build on bootstrap as well. add as hidden attribute.

* Working version support for mounting volumes on build time

Allow for mounting bind mounts at build time (and on bootstrap).

Uses docker build --build-context flags to use additional named contexts.

update both flags bootstrap and build to --mount-volumes

* fix tests for mounts

* DEV: add tests for adding volumes for docker commands

* version bump